refactor TGW modules and add a new Ref-Arch example (#59)
* Rewrite modules/transit_gateway, a breaking change.

A breaking change to become compatible with the new modules/subnet_set.

* Add modules/transit_gateway_attachment.

* Change modules/gwlb_endpoint_set to not to pass a whole module-call as an object.

Discourage the known problem: https://github.com/PaloAltoNetworks/terraform-best-practices#26-do-not-use-module-calls-as-objects

* Add examples/vmseries_combined, a new Ref-Arch example.

In the example, support the old version of vmseries module.

Have samples of init-cfg and authcodes, despite the fact it is an old way to bootstrap VM-Series.

Demo the NATGW, but EIP would work as well.

Because of overlay routing bug in panos 10.0.6 and earlier, don't use overlay routing for neither inbound or outbound. This why there are no "private" and "public" subnet sets, but just a single "data" subnet set.

Use all availability zones for GWLB, as AWS recommends. Region us-east-1 has the most zones: from "a" to "f".

It's easier later to remove zones than to add ones, considering
the subnet addressing scheme problem: Ref-Arch did not predict more than
four availability zones.