fix modules/bootstrap destroy error
Set `force_destroy` to enable the bucket to be destroyed when it
contains extra files or extra file versions beyond these handled in
Terraform.

If users upload their own files to the bucket during its life, which
happens quite often, proceed with the terraform destroy without this
error:

BucketNotEmpty: The bucket you tried to delete is not empty. You must
delete all versions in the bucket.

If users need to download any files which were put manually (authcodes
come to mind, etc.), they need to do so before destroying the bucket.
Otherwise the data will be lost.

commit-id:24ecf47c