fix avoid try() inside for_each maps
Fix a general `try()` problem illustrated by this specific example:

locals { subnets = { for k, v in var.mymap : k => try(aws_subnet.this[k], null) } }

resource "aws_route_table_association" "this" {
  for_each = var.mymap

  subnet_id      = local.subnets[each.key].id
  route_table_id = "rtb-00001234"
}

This code is a valid use of for_each. One can `terraform apply` it and
it works, initially.

The problem: Add a new entry to `mymap` and re-run `terraform apply`,
it destroys and creates again all the aws_route_table_association
resources, disrupting unnecessarily the entirety of routed traffic.
The general take is that all the dependant resources in this case are
re-created.

Removing the `try()` solves the problem, in this example only the new
aws_route_table_association is created and nothing is destroyed.

The `try()` was needed on Terraform-0.12.x, but now apparently serves no
purpose.