vpc and transit_gateway avoid try()
The `try()` can be removed to make the code more readable.
It does not impact the behavior in any noticeable way on tf version 0.13
and above.

This try() usage was previously necessary on tf version 0.12 to make the
destroy fully idempotent (to make destroy succeed even when the object was
already destroyed). This seems a 0.12-only quirk.

Many other 0.13+ codebases have worked fine without ever having such a
`try()`. The most used example:
https://github.com/PaloAltoNetworks/terraform-azurerm-vmseries-modules/blob/69afe55a596c408387c1cfabbcf598c2448ef6e6/modules/vnet/main.tf#L19-L20