vmseries_combined add input security_gwlb_service_name
Add a new optional input to the examples/vmseries_combined.
It allows to change inbound GWLB to another one easily.
The biggest reason to have it is that it allows the spoke-related items
to be easily split off from the security-vpc-related items with much
less code changes. That GWLB service name is the principal coupling of
these two halves of the example.