CI add a new git diff step
The most cases of CI failures are tf-docs, and these report back simply
the default phrase "files were modified".

This new separate step in CI will report the entire `git diff`. It would
be overly verbose for run.sh manual runs, but it's fine on GitHub, as it
folds as a separate entry in UI. It aids troubleshooting of CI fails.