transit_gateway: fix against duplicate TGWRT names (#131)
* detect re-creation of resources in tests

Assume a new testing convention: any `terraform apply` run (any of the
multiple such runs in a test case) is considered failed if any single resource
was replaced (i.e. destroyed and then re-created). Thus if you want to
explicitly test replacing a resource, split it into two test-cases:

- Can a resource be added to an existing deployment?
- Can a resource be destroyed in an existing deployment?

Why: For any of these two test-case types, it is a very rarely desired
to see any resource replaced.
But a frequent situation is that the resources are replaced in an
unexpected way and the availability/traffic suffers.

The test which replaces the entire resource is usually decomposable
into two tests - whether one can add a resource and wheter one can
destroy a resource.

Exempt the resources that do not impact data traffic.

commit-id:88a496d3

* modules/vpc fix SG tag

Correcting a mistake: the SG should get its own name, which is different
than VPC's name.

commit-id:e490098f

* transit_gateway: fix against duplicate TGWRT names

Fix for a situation when the AWS region contains two TGW-RTs with the
same Name tag. Instead of failing the import, we can easily choose the
correct TGW-RT based on the id field that we already know.

commit-id:d3b1aea9