fix(asg): Fix possible KeyError in Lambda function on terminate lifecycle.
On terminate lifecycle of the ASG lambda function invocation `instance_info['SubnetID']` may be undefined. As the parameter is not used in calling `VMSeriesInterfaceScaling.instance_terminate_action`, using ` instance_info.get('SubnetId')` (which return `None` as default) instead of accessing attribute by `[]` operatoer is an easy fix for that case.